requirements.txt
openai
python-dotenv
prompt-toolkit
regex
rich
fastapi
uvicorn
python-multipart